Bug fix
Previous code assumed that 'unsigned long' was a 64 bit type.  Since that was true on my machine, testing didn't find it...

+++ endian.c
@@ -12,7 +12,7 @@ void put_bigendian( void *target, unsigned long long value, size_t bytes ) {
     
 unsigned long long get_bigendian( const void *target, size_t bytes ) {
     const unsigned char *b = target;
-    unsigned long result = 0;
+    unsigned long long result = 0;
     int i;
 
     for (i=0; i<bytes; i++) {
